+++ CHANGELOG.md
@@ -86,6 +86,7 @@
 * ffi(nostr): add `Tag::kind_str` method ([Yuki Kishimoto])
 * ffi(nostr): impl `Display` for `Kind` ([Yuki Kishimoto])
 * js(nostr): add `JsKind::_to_string` method ([Yuki Kishimoto])
+* js(nostr): expose `from_nostr_uri` and `to_nostr_uri` for `PublicKey` and `EventId` ([Yuki Kishimoto])
 * book: add some examples ([RydalWater])
 
 ### Fixed